Changed a parameter type from a double to an int (was causing a problem with the project compiling).

+++ card-surface/CardAccount/GameAccount.cs
@@ -113,7 +113,7 @@ namespace CardAccount
         /// Balances the change.
         /// </summary>
         /// <param name="value">The value.</param>
-        public void BalanceChange(double value)
+        public void BalanceChange(int value)
         {
             balance += value;
         }
